クライアント疎通時のバグを修正

+++ api/app/Http/Controllers/_BaseController.php
@@ -112,7 +112,7 @@ abstract class _BaseController
 
         // その他の例外はシステムエラーとして扱う
         $httpStatus = $this->determineStatusCode($e);
-        $errorCode  = $e->getCode() ?: InfraErrorCode::UNKNOWN_ERROR;
+        $errorCode  = (int) ($e->getCode() ?: InfraErrorCode::UNKNOWN_ERROR);
 
         // ログに記録
         \Log::error('Exception in API request', [
